Move manipulation tests to their own file
